Mark version number string resources and not translatable
These translations are the same in all the languages and it's
better not to translate them. If they are translated, they
will be frequently outdated in translations.

(If anyone is very interested in using non-Western digits
for this, it's possibly OK for the version number, and even
then it's not recommended. And for the GitHub tag, it
shouldn't be localized at all, because it's a specific string.)